backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
python-dotenv
APScheduler
motor
pydantic
lxml